Added Violator (strain A) DOS virus to my archive
Violator (Strain A) Virus - written because the author just can't get enough of Depeche Mode and decided to write a virus in honor of their newest album at the time, Violator (what's next? Virii of Faith and Devotion?).

June 22nd, 1990 - virus goes off and is supposed to wipe the HD, just makes it unbootable, even with ```fdisk /mbr```.